Move to_cupy, nsmalltest_largets
