+++ python/cudf/cudf/tests/test_series.py
@@ -2,7 +2,6 @@
 import decimal
 import operator
 
-import cupy as cp
 import numpy as np
 import pandas as pd
 import pyarrow as pa
@@ -96,35 +95,6 @@ def test_series_contains(data, index):
     assert_eq(10 in ps, 10 in gs)
     assert_eq(True in ps, True in gs)
     assert_eq(False in ps, False in gs)
-
-
-@pytest.mark.parametrize("attr", ["nlargest", "nsmallest"])
-def test_series_nlargest_nsmallest_str_error(attr):
-    gs = cudf.Series(["a", "b", "c", "d", "e"])
-    ps = gs.to_pandas()
-
-    assert_exceptions_equal(
-        getattr(gs, attr), getattr(ps, attr), ([], {"n": 1}), ([], {"n": 1})
-    )
-
-
-def test_series_empty_dtype():
-    expected = pd.Series([])
-    actual = cudf.Series([])
-    assert_eq(expected, actual, check_dtype=True)
-
-
-@pytest.mark.parametrize("data", [None, {}, []])
-def test_series_empty_index_rangeindex(data):
-    expected = cudf.RangeIndex(0)
-    result = cudf.Series(data).index
-    assert_eq(result, expected)
-
-
-def test_series_count_invalid_param():
-    s = cudf.Series([], dtype="float64")
-    with pytest.raises(TypeError):
-        s.count(skipna=True)
 
 
 @pytest.mark.parametrize(
@@ -386,13 +356,6 @@ def test_series_duplicate_index_reindex():
     )
 
 
-@pytest.mark.parametrize("value", [1, 1.1])
-def test_nans_to_nulls_noop_copies_column(value):
-    ser1 = cudf.Series([value])
-    ser2 = ser1.nans_to_nulls()
-    assert ser1._column is not ser2._column
-
-
 @pytest.mark.parametrize(
     "type1",
     [
@@ -430,70 +393,6 @@ def test_empty_astype_always_castable(type1, type2, as_dtype, copy):
         assert ser._column is result._column
     else:
         assert ser._column is not result._column
-
-
-def test_series_dataframe_count_float():
-    gs = cudf.Series([1, 2, 3, None, np.nan, 10], nan_as_null=False)
-    ps = cudf.Series([1, 2, 3, None, np.nan, 10])
-
-    with cudf.option_context("mode.pandas_compatible", True):
-        assert_eq(ps.count(), gs.count())
-        assert_eq(ps.to_frame().count(), gs.to_frame().count())
-    with cudf.option_context("mode.pandas_compatible", False):
-        assert_eq(gs.count(), gs.to_pandas(nullable=True).count())
-        assert_eq(
-            gs.to_frame().count(),
-            gs.to_frame().to_pandas(nullable=True).count(),
-        )
-
-
-@pytest.mark.parametrize(
-    "dtype",
-    [
-        "int8",
-        "int16",
-        "int32",
-        "int64",
-        "uint8",
-        "uint16",
-        "uint32",
-        "uint64",
-        "float32",
-        "float64",
-        "bool",
-    ],
-)
-@pytest.mark.parametrize("has_nulls", [False, True])
-@pytest.mark.parametrize("use_na_value", [False, True])
-def test_series_to_cupy(dtype, has_nulls, use_na_value):
-    size = 10
-    if dtype == "bool":
-        np_data = np.array([True, False] * (size // 2), dtype=bool)
-    else:
-        np_data = np.arange(size, dtype=dtype)
-
-    if has_nulls:
-        np_data = np_data.astype("object")
-        np_data[::2] = None
-
-    sr = cudf.Series(np_data, dtype=dtype)
-
-    if not has_nulls:
-        assert_eq(sr.values, cp.asarray(sr))
-        return
-
-    if has_nulls and not use_na_value:
-        with pytest.raises(ValueError, match="Column must have no nulls"):
-            sr.to_cupy()
-        return
-
-    na_value = {
-        "bool": False,
-        "float32": 0.0,
-        "float64": 0.0,
-    }.get(dtype, 0)
-    expected = cp.asarray(sr.fillna(na_value)) if has_nulls else cp.asarray(sr)
-    assert_eq(sr.to_cupy(na_value=na_value), expected)
 
 
 def test_to_dense_array():
